Add BaseBoSampler and apply it to SkoptSampler.
