+++ optuna/samplers/__init__.py
@@ -1,6 +1,8 @@
 from optuna.samplers._search_space import intersection_search_space  # NOQA
 from optuna.samplers._search_space import IntersectionSearchSpace  # NOQA
 from optuna.samplers._base import BaseSampler  # NOQA
+from optuna.samplers._base_bo import BaseBoController  # NOQA
+from optuna.samplers._base_bo import BaseBoSampler  # NOQA
 from optuna.samplers._cmaes import CmaEsSampler  # NOQA
 from optuna.samplers._grid import GridSampler  # NOQA
 from optuna.samplers._random import RandomSampler  # NOQA
