+++ optuna/samplers/_base_bo/__init__.py
@@ -1,0 +1,2 @@
+from optuna.samplers._base_bo.controller import BaseBoController  # NOQA
+from optuna.samplers._base_bo.sampler import BaseBoSampler  # NOQA
